annotations/TraceVariables.cpp: Add local entries for PHINodes
We only support mapping this type of instruction to DILocalVariables; however,
we shouldn't ever find a phi assigned from a DIGlobalVariable because each
source of a phi's value must originate from the same DIVariable, if applicable.
This means that a phi using globals would have to source multiple copies of the
same global, which would be redundant.