GenerateEquations: Add index<-->Value mappings and populate them based on the results of the TraceVariables analysis pass
We'll need to add temporaries as we encounter them.  The most reasonable way to do this is probably
to always add them at the end, then treat them as temporaries if they occupy an index beyond the end
of the TraceVariables instance's knowledge.  To manage our equations, we can define a matrix row
class that wraps a vector (of matrix column entries) with a static size property.  Each instance
would act as if it had a number of entries exactly equal to this shared size field, regardless of
how many numbers it actually stored.  This would allow us to build a vector of these on-the-fly as
we processed the programs' instructions, then make them globally non-ragged when a client pass
requested the matrix at the end.  Alternatively, we could just have one overall matrix class that
did the fake size/raggedness hiding over a two-dimensional structure.