annotations/DimensionalAnalysis.cpp: Handle load/store and getelementptr as indirections
This adds a symbol table specifically for pointers, and only does lookups
therein where we expect a pointer register to appear.
For now, this only allows treating all accesses to a particular allocation
alike, regardless of their offset within it; however, there are two obvious ways
to extend this functionality in order to handle structured types. First, the
symbol table object could be made to point at some object characteristic of that
type of allocation (e.g. the llvm type associated with the getelementptr
instruction). Additionallly, offsets could be handled by creating another
constructor that takes an offset and manipulates the pointer that will be stored
in order to somehow incorporate it, then mapping the pointer to that shared one.