annotations/DimensionalAnalysis.cpp: Fix a severe conversion bug that produced crashes on dimens_var pass
The issue was the accidental dereference: because dimens_var doesn't define
operator*(), this caused the compiler to infer that the copy constructor---
or assignment operator, depending on where this sort of bug appears---is *not*
desired. Instead, it will use the operator DIVariable *() to coerce the original
dimens_var into a DIVariable *, dereference it as (erroneously) requested by the
programmer, then pass it to the dimens_var(const DIVariable &) constructor.
Unfortunately, this latter step will go badly when the (now one-level deep)
dimens_var instance is constructed without having a source pointer, since this
will result in passing the constructor a reference to nullptr.