annotations/DimensionalAnalysis.cpp: Fix crash by using index_mem() consistently when canonicalizing pointers
Before we were using index() in the implementations of load and store, even
though we knew we'd *populated* using mem_index(); in fact, this approach was
brain-damaged because it meant we'd make an erroneous call to insert() into the
completely wrong table even though mem_index() would've told us to skip this
step.
This fixes a crash caused by the previous commit's actually executing the
load/store setequal() codepath for the first time, and we now successfully
detect the dimensionless variable in the array-easy test program!