annotations/DimensionalAnalysis: Process inline getelementptr constant expressions
These sometimes appear directly as operands rather than as an intermediate instruction, so we need
to process them at load/store-time as well as getelementptr-time. Additionally, ConstantExprs
ironically need to be treated as non-constant in order to avoid ignoring these getelementptrs!

We now run and report dimensionless variables on the array-2d test program as well!